split: further split trace and legacy errors
