+++ weave/trace/weave_client.py
@@ -51,10 +51,6 @@ from weave.trace_server.trace_server_interface import (
     TraceServerInterface,
 )
 
-if typing.TYPE_CHECKING:
-    from ..legacy.weave import ref_base
-
-
 # Controls if objects can have refs to projects not the WeaveClient project.
 # If False, object refs with with mismatching projects will be recreated.
 # If True, use existing ref to object in other project.
@@ -904,10 +900,10 @@ class WeaveClient:
     def _remove_call_display_name(self, call: Call) -> None:
         self._set_call_display_name(call, None)
 
-    def _ref_input_to(self, ref: "ref_base.Ref") -> Sequence[Call]:
+    def _ref_input_to(self, ref: Ref) -> Sequence[Call]:
         raise NotImplementedError()
 
-    def _ref_value_input_to(self, ref: "ref_base.Ref") -> list[Call]:
+    def _ref_value_input_to(self, ref: Ref) -> list[Call]:
         raise NotImplementedError()
 
     def _ref_output_of(self, ref: ObjectRef) -> typing.Optional[Call]:
